pyproject.toml
[project]
name = "cartai"
version = "0.0.1-alpha.1"
description = "Agents for tracking your AI project lineage."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "litellm>=1.68.0",
    "python-dotenv>=1.1.0",
    "pre-commit>=4.2.0",
    "langgraph>=0.4.3",
    "aiofiles>=24.1.0",
    "fastmcp>=2.7.1",
    "mlflow>=2.22.1",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.21",
    "langchain>=0.3.25",
    "pydantic-settings>=2.9.1",
]

[project.urls]
repository = "https://www.github.com/ContrastoAI/cartai"

[project.scripts]
cartai = "cartai.deprecated.cli.main:app"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
lint = [
    "ruff<1.0.0,>=0.5.0",
    "mypy>=1.10",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
    "types-aiofiles>=24.1.0",
]
codespell = [
    "codespell<3.0.0,>=2.2.0",
]
test = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.5",
]
nb = [
    "grandalf>=0.8",
    "ipykernel>=6.29.5",
]
